// vineland/supabase/functions/api/lib/ramp/types.ts
export interface RampTxRecord {
    gatewayId: string;
    partnerTransactionId?: string;
    transactionId?: string;
    businessId?: string;
    provider: string;
    transactionType?: string;
    transactionStatus?: string;
    errorType?: string;
    errorMessage?: string;
    userId?: string;
    usdtAmount?: number;
    reaisAmount?: number;
    brCode?: string;
    hashWeb3?: string;
    endToEndId?: string;
    afterPaymentId?: string;
}